Solve all lints and warnings.
