+++ app/src/androidTest/java/com/example/starting_activity/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.startingactivity
+package com.example.starting_activity
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
