+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.example.startingactivity">
+    package="com.example.starting_activity">
 
     <application
         android:allowBackup="true"
@@ -10,49 +10,50 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
-        tools:ignore="GoogleAppIndexingWarning">
+        tools:ignore="GoogleAppIndexingWarning"
+        android:fullBackupContent="true">
         <activity
-            android:name=".StandardActivity"
+            android:name="com.example.starting_activity.StandardActivity"
             android:launchMode="standard"
-            android:parentActivityName=".MainActivity">
+            android:parentActivityName="com.example.starting_activity.MainActivity">
 
             <!-- The meta-data tag is required if you support API level 15 and lower -->
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
+                android:value="com.example.starting_activity.MainActivity" />
         </activity>
         <activity
-            android:name=".SingleTopActivity"
+            android:name="com.example.starting_activity.SingleTopActivity"
             android:launchMode="singleTop"
-            android:parentActivityName=".MainActivity">
+            android:parentActivityName="com.example.starting_activity.MainActivity">
 
             <!-- The meta-data tag is required if you support API level 15 and lower -->
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
+                android:value="com.example.starting_activity.MainActivity" />
         </activity>
         <activity
-            android:name=".SingleTaskActivity"
+            android:name="com.example.starting_activity.SingleTaskActivity"
             android:launchMode="singleTask"
-            android:parentActivityName=".MainActivity">
+            android:parentActivityName="com.example.starting_activity.MainActivity">
 
             <!-- The meta-data tag is required if you support API level 15 and lower -->
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
+                android:value="com.example.starting_activity.MainActivity" />
         </activity>
         <activity
-            android:name=".SingleInstanceActivity"
+            android:name="com.example.starting_activity.SingleInstanceActivity"
             android:launchMode="singleInstance"
-            android:parentActivityName=".MainActivity">
+            android:parentActivityName="com.example.starting_activity.MainActivity">
 
             <!-- The meta-data tag is required if you support API level 15 and lower -->
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
+                android:value="com.example.starting_activity.MainActivity" />
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name="com.example.starting_activity.MainActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
