+++ app/src/main/java/com/example/starting_activity/SingleTaskActivity.kt
@@ -1,21 +1,21 @@
-package com.example.startingactivity
+package com.example.starting_activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.TextView
 
-class SingleTopActivity : AppCompatActivity() {
+class SingleTaskActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_single_top)
+        setContentView(R.layout.activity_single_task)
 
         // Get the Intent that started this activity and extract the string
         val message = intent.getStringExtra(EXTRA_MESSAGE)
 
         // Capture the layout's TextView and set the string as its text
         findViewById<TextView>(R.id.textView).apply {
-            text = String.format(context.getString(R.string.single_top_launch_mode), message)
+            text = String.format(context.getString(R.string.single_task_launch_mode), message)
         }
     }
 }
