+++ app/src/main/res/layout/content_main.xml
@@ -9,6 +9,7 @@
     tools:showIn="@layout/activity_main">
 
     <EditText
+        android:hint="@string/main_hint"
         android:id="@+id/editText"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
@@ -17,7 +18,9 @@
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.100000024" />
+        app:layout_constraintVertical_bias="0.100000024"
+        android:inputType="textAutoComplete"
+        android:importantForAutofill="no" />
 
     <Button
         android:id="@+id/button"
